@@ -1,6 +1,6 @@
 {
   "name": "itsa-react-focuscontainer",
-  "version": "15.0.28",
+  "version": "16.0.0",
   "description": "React component: focussable container that managed the focus",
   "author": [
     {
@@ -27,21 +27,22 @@
   ],
   "main": "./lib/component.jsx",
   "engines": {
-    "node": "*"
+    "node": ">=5"
   },
   "directories": {
     "css": "./css",
     "lib": "./lib"
   },
   "dependencies": {
-    "itsa-dom": "^1.1.11",
-    "itsa-event": "^1.1.2",
-    "itsa-jsext": "^1.3.0",
-    "itsa-utils": "^1.1.0"
+    "itsa-dom": "^1.2.0",
+    "itsa-event": "^1.1.10",
+    "itsa-jsext": "^1.4.0",
+    "itsa-utils": "^1.4.0",
+    "prop-types": "^15.5.10"
   },
   "peerDependencies": {
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1"
+    "react": ">=15",
+    "react-dom": ">=15"
   },
   "devDependencies": {
     "babel-core": "^6.4.0",
@@ -52,13 +53,13 @@
     "chai": "^3.4.1",
     "chai-equal-jsx": "^1.0.9",
     "css-loader": "^0.23.1",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint": "^4.0.0",
+    "eslint-plugin-react": "^7.1.0",
     "file-loader": "^0.8.5",
-    "itsa-react-checkbox": "^15.2.16",
-    "itsa-react-input": "^15.0.17",
-    "itsa-react-maskedinput": "^15.0.17",
-    "itsa-react-textarea": "^15.0.15",
+    "itsa-react-checkbox": "^16.0.0",
+    "itsa-react-input": "^16.0.0",
+    "itsa-react-maskedinput": "^16.0.0",
+    "itsa-react-textarea": "^16.0.0",
     "istanbul": "^0.3.0",
     "jsdom": "^7.2.2",
     "jsdom-global": "^1.6.1",
@@ -67,9 +68,9 @@
     "mocha": "^1.21.4",
     "node-sass": "^3.4.2",
     "purecss": "^0.6.0",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
+    "react": "^15.6.1",
     "react-addons-test-utils": "^15.0.1",
+    "react-dom": "^15.6.1",
     "sass-loader": "^3.1.2",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.7",
