@@ -12,8 +12,7 @@
  * @since 0.0.1
 */
 
-require("itsa-jsext/lib/object");
-require("itsa-jsext/lib/array");
+require("itsa-jsext");
 require("itsa-dom");
 
 const utils = require("itsa-utils"),
@@ -32,6 +31,10 @@ const utils = require("itsa-utils"),
         KEYGEN: true,
         SELECT: true,
         TEXTAREA: true
+    },
+    PAGE_SCROLL_KEYS = {
+        38: true, // arrowUp
+        40: true // arrowDown
     },
     MAIN_CLASS = "itsa-focuscontainer",
     FOCUSCONTAINER_CLASS = "."+MAIN_CLASS,
@@ -86,6 +89,9 @@ const directionMatch = (keys, e, forbiddenKeys) => {
         }
         return match;
     });
+    if (match && PAGE_SCROLL_KEYS[keyCode]) {
+        e.preventDefault(); // no scroll of the page!
+    }
     return match;
 };
 
@@ -338,7 +344,12 @@ const defaultFnSetFocus = function(e) {
                     node.itsa_scrollIntoView(null, null, transitionTime);
                 }
                 else {
-                    node.itsa_focus && node.itsa_focus(null, null, transitionTime);
+                    if (props.scrollIntoView) {
+                        node.itsa_focus && node.itsa_focus(null, null, transitionTime);
+                    }
+                    else {
+                        node.focus();
+                    }
                 }
                 resolve(true);
             });
