@@ -4,7 +4,7 @@
 
 const React = require("react");
 const ReactDOM = require("react-dom");
-const TestUtils = require("react-addons-test-utils");
+const TestUtils = require("react-dom/test-utils");
 
 const chai = require("chai");
 const expect = chai.expect;
