@@ -4,7 +4,14 @@ module.exports = {
         path: "./examples",
         filename: "app.js" // Template based on keys in entry above
     },
-
+    externals: [
+        {
+            "../externals/react-fiber/build/packages/react/umd/react.production.min.js": "React"
+        },
+        {
+            "../externals/react-fiber/build/packages/react-dom/umd/react-dom.production.min.js": "ReactDOM"
+        }
+    ],
     module: {
         loaders: [
             {
